+++ src/migrations/Install.php
@@ -84,6 +84,7 @@ class Install extends Migration
                 'siteId' => $this->integer()->notNull(),
                 'uri' => $this->string(Blitz::$plugin->settings->maxUriLength)->notNull(),
                 'paginate' => $this->integer(),
+                'dateCached' => $this->dateTime(),
                 'expiryDate' => $this->dateTime(),
             ]);
         }